chore: encapsulation refactor for value prop block
